@@ -14,12 +14,14 @@ class For : public Instruction {
     Token::Type type;
     std::string itName;
     std::string container;
-    Block* body;
+    std::unique_ptr <Block> body;
 
 public:
     For(Token::Type type, const std::string& itName, const std::string& container);
 
-    void setBody(Block* body);
+    void setBody(std::unique_ptr <Block> body);
+
+    void execute() override;
 };
 
 
