@@ -15,8 +15,10 @@ void Variable::setPositionInContainer(const std::string& positionInContainer) {
 }
 
 Variable::Variable(const std::string& name, const std::string& positionInContainer) : name(name), positionInContainer(
-    positionInContainer), type(Type::Id) {}
+    positionInContainer), type(Type::Id) {
+    value = std::make_unique<Value> (Token::BadType, "");
+}
 
-Variable::Variable(Value* value) : value(value), type(Type::Val) {}
+Variable::Variable(std::unique_ptr <Value> value) : value(std::move(value)), type(Type::Val) {}
 
 Variable::Variable(const std::string& name) : name(name) {}
