@@ -8,6 +8,8 @@
 
 #include "../lexer/Token.h"
 #include "Value.h"
+#include <memory>
+#include <experimental/optional>
 
 class Variable {
 public:
@@ -20,12 +22,13 @@ private:
     Type type;
     std::string name; // if type == id
     std::string positionInContainer; // (optional) if type == id
-    Value* value; // if type == val
+    std::unique_ptr <Value> value; // if type == val
+//    std::experimental::optional<std::unique_ptr<Value>> value;
 
 public:
     Variable();
 
-    Variable(Value* value);
+    Variable(std::unique_ptr <Value> value);
 
     Variable(const std::string& name);
 
